LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-timeline-view-functional-notation.tentative-expected.txt


FAIL animation-timeline: view() without timeline range name assert_equals: At 0% expected "0" but got "1"
FAIL animation-timeline: view(50px) without timeline range name assert_equals: At 0% expected "0" but got "1"
FAIL animation-timeline: view(auto 50px) without timeline range name assert_equals: At 0% expected "0" but got "1"
FAIL animation-timeline: view(inline) without timeline range name assert_approx_equals: At 66.7% expected 0.33333 +/- 0.00001 but got 1
FAIL animation-timeline: view(x) without timeline range name assert_approx_equals: At 66.7% expected 0.33333 +/- 0.00001 but got 1
FAIL animation-timeline: view(y) without timeline range name assert_approx_equals: At 66.7% expected 0.33333 +/- 0.00001 but got 1
FAIL animation-timeline: view(x 50px) without timeline range name assert_equals: At 75% expected "0.25" but got "1"
FAIL animation-timeline: view(50px), view(inline 50px) without timeline range name assert_equals: At 75% inline expected "25px" but got "16px"
FAIL animation-timeline: view(inline) changes to view(inline 50px), withouttimeline range name assert_approx_equals: At 66.7% expected 0.33333 +/- 0.00001 but got 1
FAIL animation-timeline: view() assert_equals: At entry 0% expected "0" but got "1"
FAIL animation-timeline: view(50px) assert_equals: At entry 0% expected "0" but got "1"
FAIL animation-timeline: view(auto 50px) assert_equals: At entry 0% expected "0" but got "1"
FAIL animation-timeline: view(inline) assert_equals: At exit 50% expected "0.5" but got "1"
FAIL animation-timeline: view(x) assert_equals: At exit 50% expected "0.5" but got "1"
FAIL animation-timeline: view(y) assert_equals: At exit 50% expected "0.5" but got "1"
FAIL animation-timeline: view(x 50px) assert_equals: At exit 50% expected "0.5" but got "1"
FAIL animation-timeline: view(), view(inline) assert_equals: At exit 0% inline expected "10px" but got "16px"
FAIL animation-timeline: view(inline) changes to view(inline 50px) assert_equals: At exit 50% expected "0.5" but got "1"
